refactor checks, matchers, and outputs to make checks not need to use exec()
- removes use of exec() (you cannot tell if exec succeeded or failed)
- adds logging of failed checks
- startup will fail if checks.py has invalid syntax (previously it
  would silently fail checks)
- inRate and inRateKeyed uid generation is done a little differently
  (see below)

The code that created unique id's for inRate checks was replaced
with a different scheme that generates uids at run time based
on inspecting the stack and determining the file name and line number
of the code that called inRate.  There is one small caveat here in that
if there are 2 inRate checks on the same line, they will both get the
same unique ID.
